requirements.txt
Django==4.1.13
djangorestframework==3.14.0
django-allauth==0.57.2
celery==5.6.3
redis==7.4.0
stripe==15.0.0
dj-stripe==2.7.2
pytest==9.0.1
pytest-django==4.12.0
factory-boy==3.3.3
